requirements.txt
Flask~=3.1.3
SQLAlchemy~=2.0.48
python-dotenv~=1.2.2